+++ src/Client/esWMS.Client/ViewModels/WarehouseEnvironment/Location/UpdateLocationVM.cs
@@ -4,10 +4,10 @@
     {
         public string LocationId { get; set; } = null!;
         public decimal Capacity { get; set; }
-        public decimal? MaxLength { get; set; }
-        public decimal? MaxWidth { get; set; }
-        public decimal? MaxHeight { get; set; }
-        public decimal? MaxWeight { get; set; }
+        public double? MaxLength { get; set; }
+        public double? MaxWidth { get; set; }
+        public double? MaxHeight { get; set; }
+        public double? MaxWeight { get; set; }
         public string? DefaultMediaTypeId { get; set; }
     }
 }
