+++ src/Server/esWMS.Application/Functions/Documents/RwFunctions/Commands/CreateRw/CreateRwCommandHandler.cs
@@ -1,11 +1,11 @@
 ﻿using AutoMapper;
-using esWMS.Domain.Entities.Documents;
-using esWMS.Domain.Services;
 using esWMS.Application.Contracts.Persistence;
 using esWMS.Application.Contracts.Persistence.Documents;
 using esWMS.Application.Contracts.Utilities;
 using esWMS.Application.Functions.Products.Queries.GetSortedFilteredProducts;
 using esWMS.Application.Responses;
+using esWMS.Domain.Entities.Documents;
+using esWMS.Domain.Services;
 using MediatR;
 using Sieve.Models;
 
@@ -84,7 +84,7 @@ namespace esWMS.Application.Functions.Documents.RwFunctions.Commands.CreateRw
                 }
             }
 
-            Dictionary<string, int> warehouseUnitItemsQuantityToBlock = new();
+            Dictionary<string, decimal> warehouseUnitItemsQuantityToBlock = new();
 
             foreach (var documentItem in request.DocumentItems)
             {
